feat: get today habits

+++ src/components/my-habits/MyHabits.js
@@ -1,10 +1,13 @@
 import { useEffect } from "react";
 
-export default function MyHabits( { myHabit, deleteHabitFromState  } ) {
+export default function MyHabits( { myHabit, deleteHabitFromState } ) {
     const weekDays = [ "D", "S", "T", "Q", "Q", "S", "S" ];
 
     function deleteHabit() {
-        deleteHabitFromState(myHabit.id);
+        const confirmDelete = window.confirm("Tem certeza que você quer excluir este hábito?");
+        if(confirmDelete) {
+            deleteHabitFromState(myHabit.id);
+        }
     } 
 
     return (
@@ -13,7 +16,7 @@ export default function MyHabits( { myHabit, deleteHabitFromState  } ) {
                 <h3>{myHabit.name}</h3>
                 <div className="list-app">
                     {weekDays.map((weekDay, index) => (
-                        <div className={myHabit.days.indexOf(index) >= 0 ? "day-selected" : ""} id={index}>{weekDay}</div> 
+                        <div key={index} className={myHabit.days.indexOf(index) >= 0 ? "day-selected" : ""}>{weekDay}</div> 
                     ))}
                 </div>
             </div>
